# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CodeCraft API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: languages
    description: Programming languages
  - name: courses
    description: Course management
  - name: lessons
    description: Lesson content
  - name: progress
    description: User progress tracking
  - name: quiz
    description: Quiz and assessments
  - name: code
    description: Code execution
  - name: leaderboard
    description: Leaderboard
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /languages:
    get:
      operationId: listLanguages
      tags: [languages]
      summary: List all programming languages
      responses:
        "200":
          description: List of languages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Language"

  /languages/{slug}/courses:
    get:
      operationId: listCoursesByLanguage
      tags: [languages]
      summary: List courses for a language
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Courses for the language
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"

  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course details with lessons
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course with lessons
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
        "404":
          description: Not found

  /lessons/{id}:
    get:
      operationId: getLesson
      tags: [lessons]
      summary: Get a single lesson
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
        "404":
          description: Not found

  /progress:
    get:
      operationId: getUserProgress
      tags: [progress]
      summary: Get authenticated user's progress
      responses:
        "200":
          description: User progress
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProgress"
        "401":
          description: Unauthorized
    post:
      operationId: markLessonComplete
      tags: [progress]
      summary: Mark a lesson as complete
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressInput"
      responses:
        "200":
          description: Progress updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProgress"
        "401":
          description: Unauthorized

  /progress/summary:
    get:
      operationId: getProgressSummary
      tags: [progress]
      summary: Get summary of user progress per language
      responses:
        "200":
          description: Progress summary
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgressSummary"
        "401":
          description: Unauthorized

  /quiz/course/{courseId}:
    get:
      operationId: getQuizByCourse
      tags: [quiz]
      summary: Get quiz for a course
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
        "404":
          description: Not found

  /quiz/attempts:
    post:
      operationId: submitQuizAttempt
      tags: [quiz]
      summary: Submit a solo quiz attempt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizAttemptInput"
      responses:
        "201":
          description: Attempt result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizAttempt"
        "401":
          description: Unauthorized

  /quiz/attempts/me:
    get:
      operationId: getMyQuizAttempts
      tags: [quiz]
      summary: Get authenticated user's quiz attempts
      responses:
        "200":
          description: List of quiz attempts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuizAttempt"
        "401":
          description: Unauthorized

  /quiz/sessions:
    post:
      operationId: createQuizSession
      tags: [quiz]
      summary: Create a multiplayer quiz session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSessionInput"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizSession"
        "401":
          description: Unauthorized

  /quiz/sessions/{code}:
    get:
      operationId: getQuizSession
      tags: [quiz]
      summary: Get a multiplayer quiz session by code
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizSession"
        "404":
          description: Not found

  /quiz/sessions/{code}/join:
    post:
      operationId: joinQuizSession
      tags: [quiz]
      summary: Join a multiplayer quiz session
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Joined session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizSession"
        "404":
          description: Not found
        "401":
          description: Unauthorized

  /quiz/sessions/{code}/answer:
    post:
      operationId: submitSessionAnswer
      tags: [quiz]
      summary: Submit an answer in a multiplayer session
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionAnswerInput"
      responses:
        "200":
          description: Answer recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizSession"
        "401":
          description: Unauthorized

  /code/run:
    post:
      operationId: runCode
      tags: [code]
      summary: Execute code and return output
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeRunInput"
      responses:
        "200":
          description: Code execution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeRunResult"

  /leaderboard:
    get:
      operationId: getLeaderboard
      tags: [leaderboard]
      summary: Get top users by XP/score
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
        - in: query
          name: languageSlug
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get platform-wide stats (admin only)
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden

  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users with stats (admin only)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: search
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Paginated users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Language:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        color:
          type: string
        icon:
          type: string
        courseCount:
          type: integer
      required:
        - id
        - name
        - slug
        - description
        - color
        - icon
        - courseCount

    Course:
      type: object
      properties:
        id:
          type: integer
        languageId:
          type: integer
        title:
          type: string
        description:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        lessonCount:
          type: integer
        xpReward:
          type: integer
      required:
        - id
        - languageId
        - title
        - description
        - level
        - lessonCount
        - xpReward

    Lesson:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        content:
          type: string
        codeExample:
          type: ["string", "null"]
        language:
          type: string
        order:
          type: integer
        xpReward:
          type: integer
      required:
        - id
        - courseId
        - title
        - content
        - language
        - order
        - xpReward

    CourseDetail:
      type: object
      properties:
        id:
          type: integer
        languageId:
          type: integer
        title:
          type: string
        description:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        xpReward:
          type: integer
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"
      required:
        - id
        - languageId
        - title
        - description
        - level
        - xpReward
        - lessons

    UserProgress:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        lessonId:
          type: integer
        completed:
          type: boolean
        completedAt:
          type: ["string", "null"]
      required:
        - id
        - userId
        - lessonId
        - completed

    ProgressInput:
      type: object
      properties:
        lessonId:
          type: integer
      required:
        - lessonId

    ProgressSummary:
      type: object
      properties:
        languageId:
          type: integer
        languageName:
          type: string
        languageSlug:
          type: string
        totalLessons:
          type: integer
        completedLessons:
          type: integer
        xpEarned:
          type: integer
      required:
        - languageId
        - languageName
        - languageSlug
        - totalLessons
        - completedLessons
        - xpEarned

    Quiz:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
      required:
        - id
        - courseId
        - title
        - questions

    QuizQuestion:
      type: object
      properties:
        id:
          type: integer
        question:
          type: string
        options:
          type: array
          items:
            type: string
        correctIndex:
          type: integer
      required:
        - id
        - question
        - options
        - correctIndex

    QuizAttemptInput:
      type: object
      properties:
        quizId:
          type: integer
        answers:
          type: array
          items:
            type: integer
      required:
        - quizId
        - answers

    QuizAttempt:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        quizId:
          type: integer
        score:
          type: integer
        totalQuestions:
          type: integer
        xpAwarded:
          type: integer
        completedAt:
          type: string
      required:
        - id
        - userId
        - quizId
        - score
        - totalQuestions
        - xpAwarded
        - completedAt

    QuizSessionInput:
      type: object
      properties:
        quizId:
          type: integer
      required:
        - quizId

    QuizSession:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        quizId:
          type: integer
        status:
          type: string
          enum: [waiting, active, finished]
        hostUserId:
          type: string
        participants:
          type: array
          items:
            $ref: "#/components/schemas/SessionParticipant"
        currentQuestion:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - code
        - quizId
        - status
        - hostUserId
        - participants
        - createdAt

    SessionParticipant:
      type: object
      properties:
        userId:
          type: string
        displayName:
          type: string
        score:
          type: integer
        answeredCount:
          type: integer
      required:
        - userId
        - displayName
        - score
        - answeredCount

    SessionAnswerInput:
      type: object
      properties:
        questionId:
          type: integer
        answerIndex:
          type: integer
      required:
        - questionId
        - answerIndex

    CodeRunInput:
      type: object
      properties:
        language:
          type: string
          enum: [html, css, javascript, java, c, python]
        code:
          type: string
      required:
        - language
        - code

    CodeRunResult:
      type: object
      properties:
        output:
          type: string
        error:
          type: ["string", "null"]
        runtime:
          type: ["number", "null"]
      required:
        - output

    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        userId:
          type: string
        displayName:
          type: string
        xp:
          type: integer
        completedCourses:
          type: integer
        quizzesPassed:
          type: integer
      required:
        - rank
        - userId
        - displayName
        - xp
        - completedCourses
        - quizzesPassed

    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        activeUsersToday:
          type: integer
        activeUsersWeek:
          type: integer
        totalLessonsCompleted:
          type: integer
        totalQuizAttempts:
          type: integer
        popularLanguage:
          type: string
        languageBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/LanguageStat"
      required:
        - totalUsers
        - activeUsersToday
        - activeUsersWeek
        - totalLessonsCompleted
        - totalQuizAttempts
        - popularLanguage
        - languageBreakdown

    LanguageStat:
      type: object
      properties:
        languageName:
          type: string
        userCount:
          type: integer
        lessonsCompleted:
          type: integer
      required:
        - languageName
        - userCount
        - lessonsCompleted

    AdminUserList:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - users
        - total
        - page
        - limit

    AdminUser:
      type: object
      properties:
        userId:
          type: string
        displayName:
          type: string
        email:
          type: string
        xp:
          type: integer
        lessonsCompleted:
          type: integer
        quizzesPassed:
          type: integer
        joinedAt:
          type: string
        lastActive:
          type: ["string", "null"]
      required:
        - userId
        - displayName
        - email
        - xp
        - lessonsCompleted
        - quizzesPassed
        - joinedAt