# lib/api-zod/src/generated/types/course.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CodeCraft API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CourseLevel } from './courseLevel';

export interface Course {
  id: number;
  languageId: number;
  title: string;
  description: string;
  level: CourseLevel;
  lessonCount: number;
  xpReward: number;
}