# lib/api-zod/src/generated/types/languageStat.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CodeCraft API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LanguageStat {
  languageName: string;
  userCount: number;
  lessonsCompleted: number;
}