# lib/api-zod/src/generated/types/lesson.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CodeCraft API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Lesson {
  id: number;
  courseId: number;
  title: string;
  content: string;
  /** @nullable */
  codeExample?: string | null;
  language: string;
  order: number;
  xpReward: number;
}