# lib/api-zod/src/generated/types/progressSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CodeCraft API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProgressSummary {
  languageId: number;
  languageName: string;
  languageSlug: string;
  totalLessons: number;
  completedLessons: number;
  xpEarned: number;
}